Refactor command structure to make the command tree usable by other mains
This is the first step in refactor gnmic to allow for a minimum cli
which can be used as a lighterweight implementation for either on box
or in embedded or low dependency environments.
