+++ pkg/cmd/prompt_test.go
